requirements.txt
yt-dlp>=2024.0.0
rich>=13.0.0
InquirerPy>=0.3.4
mutagen>=1.47.0
pyyaml>=6.0
pydantic>=2.0.0